refactor(timeline): split REST helpers into smaller modules
